Move collection serializers to own file
